+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/MagicController.java
@@ -2438,6 +2438,7 @@ public class MagicController implements MageController {
 		// Parse wand settings
 		Wand.DefaultUpgradeMaterial = ConfigurationUtils.getMaterial(properties, "wand_upgrade_item", Wand.DefaultUpgradeMaterial);
         Wand.SpellGlow = properties.getBoolean("spell_glow", Wand.SpellGlow);
+        Wand.LiveHotbarSkills = properties.getBoolean("live_hotbar_skills", Wand.LiveHotbarSkills);
         Wand.LiveHotbar = properties.getBoolean("live_hotbar", Wand.LiveHotbar);
         Wand.LiveHotbarCooldown = properties.getBoolean("live_hotbar_cooldown", Wand.LiveHotbar);
         Wand.BrushGlow = properties.getBoolean("brush_glow", Wand.BrushGlow);