+++ Magic/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add ModifyMana action
  - Fix/Improve PVPManager integration (don't target players with PVP disabled)
  - Added "example: skulls" option to easily re-enable skull-based icons
+ - Add live_hotbar_skills option to allow skills to have a cooldown timer
 
 ## 6.3.3